Progect biology/requirements.txt
PyQt 5
sys
random